# internal/restic: Custom ID.MarshalJSON